+++ doc/OnlineDocs/contributed_packages/index.rst
@@ -30,7 +30,6 @@ Contributed packages distributed with Pyomo:
    pyros.rst
    sensitivity_toolbox.rst
    trustregion.rst
-   mis.rst
 
 Contributed Pyomo interfaces to other packages:
