+++ pyomo/contrib/iis/tests/trivial_mis.py
